Cambios en servicios y estructura

+++ backend/src/services/doctors.service.ts
@@ -26,6 +26,13 @@ export const getDoctorById = async (id: string): Promise<Doctor> => {
 }
 
 export const createDoctor = async (dto: CreateDoctorDto): Promise<Doctor> => {
+
+    const existingDoctor = await doctorRepository.findOne({
+        where: { name: dto.name }
+    })
+
+    if (existingDoctor) throw new AppError("Doctor already exists", 400)
+
     return await doctorRepository.save(
         doctorRepository.create(dto)
     )
@@ -37,6 +44,14 @@ export const updateDoctor = async (id: string, dto: UpdateDoctorDto): Promise<Do
 
     if (!existingDoctor) throw new AppError("Doctor not found", 404)
 
+    if (dto.name) {
+        const doctorWithSameName = await doctorRepository.findOne({
+            where: { name: dto.name }
+        })
+
+        if (doctorWithSameName && doctorWithSameName.id !== id) throw new AppError("Doctor already exists", 400)
+    }
+
     Object.assign(existingDoctor, dto)
 
     return await doctorRepository.save(existingDoctor)
@@ -46,6 +61,10 @@ export const deleteDoctor = async (id: string): Promise<void> => {
     const existingDoctor = await doctorRepository.findOneBy({ id })
 
     if (!existingDoctor) throw new AppError('Doctor not found', 404)
+
+    const appointments = await appointmentsRepository.count({ where: { doctor: { id } } })
+
+    if(appointments > 0) throw new AppError('Cannot delete doctor with associated appointments', 400)
 
     await doctorRepository.remove(existingDoctor)
 }
